+++ CHANGELOG.md
@@ -13,7 +13,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-* `cut_rows` and `cut_objects` functions to cut off scan key and primary key values that were merged to the select/pairs partial result.
+* `cut_rows` and `cut_objects` functions to cut off scan key and
+  primary key values that were merged to the select/pairs partial result.
+* Functions ``stop()`` for the roles ``crud-storage`` and ``crud-router``.
 
 ## [0.6.0] - 2021-03-29